✨ feat(icm42688p:修改read吐出的格式为温度float其他未int16)

+++ bos/drivers/inc/b_driver_cmd.h
@@ -102,9 +102,9 @@ typedef struct
 {
     float temperature;
 
-    float acc_arr[3];
+    int16_t acc_arr[3];
 
-    float gyro_arr[3];
+    int16_t gyro_arr[3];
 } bICM42688P_6Axis_t;
 
 typedef struct
